+++ DeviceDiagnostics/DeviceDiagnostics.cpp
@@ -36,7 +36,7 @@
 
 #define API_VERSION_NUMBER_MAJOR 1
 #define API_VERSION_NUMBER_MINOR 1
-#define API_VERSION_NUMBER_PATCH 2
+#define API_VERSION_NUMBER_PATCH 3
 
 enum SysSrv_ErrorCode {
     SysSrv_FileNotPresent,