+++ DeviceIdentification/CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to this RDK Service will be documented in this file.
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 
+## [1.0.6] - 2024-05-28
+### Fixed
+- Added methods for retrieving Serial number, Chip Id and Firmware version for Broadcom Devices.
+
 ## [1.0.5] - 2024-01-10
 ### Fixed
 - Add chipset name in DeviceIdentification plugin