+++ HdmiCec_2/CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to this RDK Service will be documented in this file.
 * Changes in CHANGELOG should be updated when commits are added to the main or release branches. There should be one CHANGELOG entry per JIRA Ticket. This is not enforced on sprint branches since there could be multiple changes for the same JIRA ticket during development. 
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
+## [1.0.14] - 2024-05-25
+### Added
+- Make plugin autostart configurable from recipe
+
 ## [1.0.13] - 2023-01-02
 ### Changed
 - Updated  HdmiCec_2 conf.in based on Modern config generator